refactor common code into `article` package, fix #94
- add article package, move yarden2annot module into it
- fix entry point in setup.py for crowsetta format 'yarden'
- add src/article/eval/metrics with functions from scripts
- add src/article/eval/eval.py with functions from scripts
- rewrite make_error scripts to use article package
- add src/article/plot with error_curve.py module
- update install instructions in README.md
  + to explain user needs to specify `[article]`
    when installing project to re-run / reproduce analyses
