+++ src/article/plot/__init__.py
@@ -1,0 +1,1 @@
+from . import error_curve
